+++ nuvlaedge/common/constants.go
@@ -1,5 +1,7 @@
 package common
 
+import log "github.com/sirupsen/logrus"
+
 // Nuvla Endpoint constant configuration
 const (
 	NuvlaEndPoint  string = "https://nuvla.io"
@@ -14,13 +16,13 @@ const (
 	SessionEndpoint = "/api/session"
 )
 
-// BasePath NuvlaEdge local configuration path constants
-// ConfPath
-// NuvlaEdgeLocalDB
 const (
-	BasePath         string = "/etc/nuvlaedge/"
-	ConfPath                = BasePath + "config/"
-	NuvlaEdgeLocalDB        = BasePath + ".local/"
+	// BasePath NuvlaEdge local configuration path constants
+	BasePath string = "/etc/nuvlaedge/"
+	// ConfPath Location to NuvlaEdge configuration files locally
+	ConfPath = BasePath + "config/"
+	// NuvlaEdgeLocalDB Local database path
+	NuvlaEdgeLocalDB = BasePath + ".local/"
 )
 
 // NuvlaEdgeUserConfig
@@ -30,3 +32,22 @@ const (
 )
 
 // TODO: Release usage of pathlib
+
+// BaseImageName common NuvlaEdge image
+// Image Constants
+const (
+	BaseImageName string = "alpine:3.18"
+)
+
+// LogLevel Global log level propagator
+var LogLevel log.Level = log.InfoLevel
+
+func SetLogLevel(level string) {
+	l, err := log.ParseLevel(level)
+	if err != nil {
+		log.Warnf("Invalid log level: %s", level)
+		LogLevel = log.InfoLevel
+	} else {
+		LogLevel = l
+	}
+}